Adding Selected Bar Style
